@@ -3,10 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "node scripts/generate-document-index.js && next dev",
+    "dev": "node scripts/generate-document-index.mjs && next dev",
     "prebuild": "find . -name \".DS_Store\" -type f -delete || true",
-    "build": "node scripts/generate-document-index.js && next build",
-    "generate-docs": "node scripts/generate-document-index.js",
+    "build": "node scripts/generate-document-index.mjs && next build",
+    "generate-docs": "node scripts/generate-document-index.mjs",
     "start": "next start",
     "lint": "next lint",
     "sitemap": "next-sitemap"
@@ -23,7 +23,7 @@
     "express": "^4.21.2",
     "glob": "^10.3.12",
     "js-yaml": "^4.1.0",
-    "next": "^15.1.5",
+    "next": "^15.1.11",
     "next-compose-plugins": "^2.2.1",
     "next-sitemap": "^4.2.3",
     "node-fetch": "^3.3.2",
@@ -31,15 +31,13 @@
     "react": "^18",
     "react-dom": "^18",
     "react-markdown": "^9.0.1",
-    "serve": "^14.2.4",
     "sharp": "^0.33.3",
-    "xlsx": "^0.18.5",
     "yaml": "^2.7.0"
   },
   "devDependencies": {
     "autoprefixer": "^10.0.1",
-    "eslint": "^8",
-    "eslint-config-next": "14.1.4",
+    "eslint": "^9.39.2",
+    "eslint-config-next": "^16.1.1",
     "postcss": "^8"
   }
 }
