@@ -1,5 +1,5 @@
-const fs = require('fs')
-const fetch = (...args) => import('node-fetch').then(({ default: fetch }) => fetch(...args));
+import fs from 'fs';
+import fetch from 'node-fetch';
 
 const writeJson = (path, data) => {
   fs.writeFile(path, JSON.stringify(data, null, 2), err => {
